cloud_function/upload_to_gcs/requirements.txt
google-cloud-storage==1.35.1